handler/requirements.txt
requests
redis
sqlalchemy
paho-mqtt==1.4.0
